python_assistant/requirements.txt
groq
SpeechRecognition
pyaudio
edge-tts
python-dotenv
pygame
colorama
elevenlabs
